Aug 28, 2023, 7:53 AM

+++ ch_08/q1.java
@@ -5,11 +5,11 @@ import java.util.Scanner;
 public class q1 {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        int[][] mat = new int[3][4];
+        double[][] mat = new double[3][4];
         System.out.println("Enter a 3-by-4 matrix row by row: ");
         for (int i = 0; i < 3; i++) {
             for (int j =0; j < 4; j++) {
-                mat[i][j] = sc.nextInt();
+                mat[i][j] = sc.nextDouble();
             }
         }
